opt: fix overflow causing zero row-count
We were subtracting two ints that could overflow and then casting the
result to a float64. There's an easy solution to avoid the overflow:
just cast each integer to a float64 before performing the subtraction.

I will backport this to 19.1 and 2.1.

Release note (bug fix): Previously, due to a bug when estimating result
set sizes in the optimizer, queries involving int ranges that were very
large could result in poor plans being generated.

+++ pkg/sql/opt/memo/statistics_builder.go
@@ -2527,9 +2527,9 @@ func (sb *statisticsBuilder) updateDistinctCountsFromConstraint(
 					// should be the case for integer valued columns (due to normalization
 					// by constraint.PreferInclusive).
 					if c.Columns.Get(col).Ascending() {
-						distinctCount += float64(end - start)
+						distinctCount += float64(end) - float64(start)
 					} else {
-						distinctCount += float64(start - end)
+						distinctCount += float64(start) - float64(end)
 					}
 				} else {
 					// We can't determine the distinct count for this column. For example,
